wip: split InstanceController out to have InstanceModules
