+++ lib/Data/ImportExport.js
@@ -403,7 +403,7 @@ class DataImportExport extends CoreBase {
 				if (instanceId === 'internal' || instanceId === 'bitfocus-companion') continue
 
 				clientObject.instances[instanceId] = {
-					instance_type: this.instance.verifyInstanceTypeIsCurrent(instance.instance_type),
+					instance_type: this.instance.modules.verifyInstanceTypeIsCurrent(instance.instance_type),
 					label: instance.label,
 				}
 			}
@@ -693,7 +693,7 @@ class DataImportExport extends CoreBase {
 				}
 			} else {
 				// Create a new instance
-				const instance_type = this.instance.verifyInstanceTypeIsCurrent(obj.instance_type)
+				const instance_type = this.instance.modules.verifyInstanceTypeIsCurrent(obj.instance_type)
 				const newId = this.instance.addInstance({ type: instance_type }, true)
 				console.log('created', instance_type, newId)
 				if (newId) {
