+++ lib/Instance/Controller.js
@@ -15,27 +15,22 @@
  *
  */
 
-import fs from 'fs-extra'
 import { nanoid } from 'nanoid'
-import { isPackaged, sleep } from '../Resources/Util.js'
+import { sleep } from '../Resources/Util.js'
 import CoreBase from '../Core/Base.js'
 import InstanceDefinitions from './Definitions.js'
 import InstanceVariable from './Variable.js'
-import path from 'path'
 import ModuleHost, { ConnectionDebugLogRoom } from './Host.js'
 import InstanceStatus from './Status.js'
-import { validateManifest } from '@companion-module/base'
-import { fileURLToPath } from 'url'
 import { cloneDeep } from 'lodash-es'
 import jsonPatch from 'fast-json-patch'
 import { isLabelValid, makeLabelSafe } from '../Shared/Label.js'
+import InstanceModules from './Modules.js'
 
 const InstancesRoom = 'instances'
-const ModulesRoom = 'modules'
 
 class Instance extends CoreBase {
 	#lastClientJson = null
-	#lastModulesJson = null
 
 	constructor(registry) {
 		super(registry, 'instance', 'Instance/Controller')
@@ -44,11 +39,7 @@ class Instance extends CoreBase {
 		this.definitions = new InstanceDefinitions(registry)
 		this.status = new InstanceStatus(registry)
 		this.moduleHost = new ModuleHost(registry, this.status)
-
-		/** Object of the known modules that can be loaded */
-		this.known_modules = {}
-		/** Sometimes modules get renamed/merged. This lets that happen */
-		this.module_renames = {}
+		this.modules = new InstanceModules(registry)
 
 		this.store = {
 			db: {},
@@ -58,24 +49,6 @@ class Instance extends CoreBase {
 
 		// Prepare for clients already
 		this.commitChanges()
-
-		this.registry.api_router.get('/help/module/:module_id/*', (req, res, next) => {
-			const module_id = req.params.module_id.replace(/\.\.+/g, '')
-			const file = req.params[0].replace(/\.\.+/g, '')
-
-			const moduleInfo = this.known_modules[module_id]
-			if (moduleInfo && moduleInfo.helpPath && moduleInfo.basePath) {
-				const fullpath = path.join(moduleInfo.basePath, 'companion', file)
-				if (file.match(/\.(jpe?g|gif|png|pdf)$/) && fs.existsSync(fullpath)) {
-					// Send the file, then stop
-					res.sendFile(fullpath)
-					return
-				}
-			}
-
-			// Try next handler
-			next()
-		})
 	}
 
 	getAllInstanceIds() {
@@ -109,129 +82,26 @@ class Instance extends CoreBase {
 	async initInstances(extraModulePath) {
 		this.logger.silly('instance_init', this.store.db)
 
-		const rootPath = isPackaged() ? path.join(__dirname, '.') : fileURLToPath(new URL('../../', import.meta.url))
-
-		const searchDirs = [
-			// Paths to look for modules, lowest to highest priority
-			path.resolve(path.join(rootPath, 'bundled-modules')),
-		]
-
-		const legacyCandidates = await this.#loadInfoForModulesInDir(path.join(rootPath, '/module-legacy/manifests'), false)
-
-		// Start with 'legacy' candidates
-		for (const candidate of legacyCandidates) {
-			candidate.display.isLegacy = true
-			this.known_modules[candidate.manifest.id] = candidate
-		}
-
-		// Load modules from other folders in order of priority
-		for (const searchDir of searchDirs) {
-			const candidates = await this.#loadInfoForModulesInDir(searchDir, false)
-			for (const candidate of candidates) {
-				// Replace any existing candidate
-				this.known_modules[candidate.manifest.id] = candidate
-			}
-		}
-
-		if (extraModulePath) {
-			this.logger.info(`Looking for extra modules in: ${extraModulePath}`)
-			const candidates = await this.#loadInfoForModulesInDir(extraModulePath, true)
-			for (const candidate of candidates) {
-				// Replace any existing candidate
-				this.known_modules[candidate.manifest.id] = {
-					...candidate,
-					isOverride: true,
-				}
-			}
-
-			this.logger.info(`Found ${candidates.length} extra modules`)
-		}
-
-		// Figure out the redirects. We do this afterwards, to ensure we avoid collisions and stuff
-		for (const id of Object.keys(this.known_modules).sort()) {
-			const moduleInfo = this.known_modules[id]
-			if (moduleInfo && Array.isArray(moduleInfo.manifest.legacyIds)) {
-				if (moduleInfo.display.isLegacy) {
-					// Handle legacy modules differently. They should never replace a new style one
-					for (const legacyId of moduleInfo.manifest.legacyIds) {
-						const otherInfo = this.known_modules[legacyId]
-						if (!otherInfo || otherInfo.display.isLegacy) {
-							// Other is not known or is legacy
-							this.module_renames[legacyId] = id
-							delete this.known_modules[legacyId]
-						}
-					}
-				} else {
-					// These should replace anything
-					for (const legacyId of moduleInfo.manifest.legacyIds) {
-						this.module_renames[legacyId] = id
-						delete this.known_modules[legacyId]
-					}
-				}
-			}
-		}
-
-		// Log the loaded modules
-		for (const id of Object.keys(this.known_modules).sort()) {
-			const moduleInfo = this.known_modules[id]
-			if (moduleInfo.isOverride) {
-				this.logger.info(
-					`${moduleInfo.display.id}@${moduleInfo.display.version}: ${moduleInfo.display.name} (Overridden${
-						moduleInfo.isPackaged ? ' & Packaged' : ''
-					})`
-				)
-			} else {
-				this.logger.debug(`${moduleInfo.display.id}@${moduleInfo.display.version}: ${moduleInfo.display.name}`)
-			}
-		}
+		await this.modules.initInstances(extraModulePath)
 
 		for (const id in this.store.db) {
 			this.activate_module(id, false)
 		}
 	}
 
-	async reloadExtraModule(fullpath) {
-		this.logger.info(`Attempting to reload module in: ${fullpath}`)
-
-		const reloadedModule = await this.#loadInfoForModule(fullpath, true)
-		if (reloadedModule) {
-			this.logger.info(
-				`Found new module "${reloadedModule.display.id}" v${reloadedModule.display.version} in: ${fullpath}`
-			)
-
-			// Replace any existing module
-			this.known_modules[reloadedModule.manifest.id] = {
-				...reloadedModule,
-				isOverride: true,
+	reloadUsesOfModule(module_id) {
+		// restart usages of this module
+		let reloadLabels = []
+		for (const [id, instance_config] of Object.entries(this.store.db)) {
+			if (instance_config.instance_type === module_id) {
+				reloadLabels.push(instance_config.label)
+				// Restart it
+				this.enableDisableInstance(id, false)
+				this.enableDisableInstance(id, true)
 			}
-
-			const newJson = cloneDeep(this.getModulesJson())
-
-			// Now broadcast to any interested clients
-			if (this.io.countRoomMembers(ModulesRoom) > 0) {
-				const patch = jsonPatch.compare(this.#lastModulesJson || {}, newJson || {})
-				if (patch.length > 0) {
-					this.io.emitToRoom(ModulesRoom, `modules:patch`, patch)
-				}
-			}
-
-			this.#lastModulesJson = newJson
-
-			// restart usages of this module
-			let reloadLabels = []
-			for (const [id, instance_config] of Object.entries(this.store.db)) {
-				if (instance_config.instance_type === reloadedModule.manifest.id) {
-					reloadLabels.push(instance_config.label)
-					// Restart it
-					this.enableDisableInstance(id, false)
-					this.enableDisableInstance(id, true)
-				}
-			}
-
-			this.logger.info(`Reloading ${reloadLabels.length} connections: ${reloadLabels.join(', ')}`)
-		} else {
-			this.logger.info(`Failed to find module in: ${fullpath}`)
 		}
+
+		this.logger.info(`Reloading ${reloadLabels.length} connections: ${reloadLabels.join(', ')}`)
 	}
 
 	setInstanceLabelAndConfig(id, newLabel, config, skip_notify_instance) {
@@ -311,7 +181,7 @@ class Instance extends CoreBase {
 					.filter((n) => typeof n === 'number')
 			) || 0
 
-		const moduleInfo = this.known_modules[module]
+		const moduleInfo = this.modules.known_modules[module]
 		if (moduleInfo) {
 			let id = nanoid()
 
@@ -419,15 +289,6 @@ class Instance extends CoreBase {
 	}
 
 	/**
-	 * Checks whether an instance_type has been renamed
-	 * @param {string} instance_type
-	 * @returns {string} the instance_type that should be used (often the provided parameter)
-	 */
-	verifyInstanceTypeIsCurrent(instance_type) {
-		return this.module_renames[instance_type] || instance_type
-	}
-
-	/**
 	 * Get information for the metrics system about the current instances
 	 */
 	getInstancesMetrics() {
@@ -520,7 +381,7 @@ class Instance extends CoreBase {
 		const config = this.store.db[id]
 		if (!config) throw new Error('Cannot activate unknown module')
 
-		config.instance_type = this.verifyInstanceTypeIsCurrent(config.instance_type)
+		config.instance_type = this.modules.verifyInstanceTypeIsCurrent(config.instance_type)
 
 		if (config.enabled === false) {
 			this.logger.silly("Won't load disabled module " + id + ' (' + config.instance_type + ')')
@@ -534,7 +395,7 @@ class Instance extends CoreBase {
 			this.setInstanceLabelAndConfig(id, safeLabel, null, true)
 		}
 
-		const moduleInfo = this.known_modules[config.instance_type]
+		const moduleInfo = this.modules.known_modules[config.instance_type]
 		if (!moduleInfo) {
 			this.logger.error('Configured instance ' + config.instance_type + ' could not be loaded, unknown module')
 		} else {
@@ -553,6 +414,7 @@ class Instance extends CoreBase {
 		this.variable.clientConnect(client)
 		this.definitions.clientConnect(client)
 		this.status.clientConnect(client)
+		this.modules.clientConnect(client)
 
 		client.onPromise('instances:subscribe', () => {
 			client.join(InstancesRoom)
@@ -561,16 +423,6 @@ class Instance extends CoreBase {
 		})
 		client.onPromise('instances:unsubscribe', () => {
 			client.leave(InstancesRoom)
-		})
-
-		client.onPromise('modules:subscribe', () => {
-			client.join(ModulesRoom)
-
-			return this.#lastModulesJson || this.getModulesJson()
-		})
-
-		client.onPromise('modules:unsubscribe', () => {
-			client.leave(ModulesRoom)
 		})
 
 		client.onPromise('instances:edit', async (id) => {
@@ -635,21 +487,6 @@ class Instance extends CoreBase {
 			return id
 		})
 
-		client.onPromise('instances:get-help', async (module_id) => {
-			try {
-				const res = await this.getHelpForModule(module_id)
-				if (res) {
-					return [null, res]
-				} else {
-					return ['nofile', null]
-				}
-			} catch (err) {
-				this.logger.silly(`Error loading help for ${module_id}`)
-				this.logger.silly(err)
-				return ['nofile', null]
-			}
-		})
-
 		client.onPromise('instances:set-order', async (instanceIds) => {
 			if (!Array.isArray(instanceIds)) throw new Error('Expected array of ids')
 
@@ -689,16 +526,6 @@ class Instance extends CoreBase {
 		})
 	}
 
-	getModulesJson() {
-		const result = {}
-
-		for (const [id, module] of Object.entries(this.known_modules)) {
-			result[id] = module.display
-		}
-
-		return result
-	}
-
 	getClientJson() {
 		const result = {}
 
@@ -720,116 +547,6 @@ class Instance extends CoreBase {
 		}
 
 		return result
-	}
-
-	/**
-	 * Load the help markdown file for a specified module_id
-	 * @access public
-	 * @param {string} module_id
-	 */
-	async getHelpForModule(module_id) {
-		const moduleInfo = this.known_modules[module_id]
-		if (moduleInfo && moduleInfo.helpPath) {
-			const stats = await fs.stat(moduleInfo.helpPath)
-			if (stats.isFile()) {
-				const data = await fs.readFile(moduleInfo.helpPath)
-				return {
-					markdown: data.toString(),
-					baseUrl: `/int/help/module/${module_id}/`,
-				}
-			} else {
-				this.logger.silly(`Error loading help for ${module_id}`, moduleInfo.helpPath)
-				this.logger.silly('Not a file')
-				return undefined
-			}
-		} else {
-			return undefined
-		}
-	}
-
-	/**
-	 * Load information about all modules in a directory
-	 * @access private
-	 * @param {string} searchDir - Path to search for modules
-	 * @param {boolean} checkForPackaged - Whether to check for a packaged version
-	 */
-	async #loadInfoForModulesInDir(searchDir, checkForPackaged) {
-		if (await fs.pathExists(searchDir)) {
-			const candidates = await fs.readdir(searchDir)
-
-			const ps = []
-
-			for (const candidate of candidates) {
-				const candidatePath = path.join(searchDir, candidate)
-				ps.push(this.#loadInfoForModule(candidatePath, checkForPackaged))
-			}
-
-			const res = await Promise.all(ps)
-			return res.filter((v) => !!v)
-		} else {
-			return []
-		}
-	}
-
-	/**
-	 * Load information about a module
-	 * @access private
-	 * @param {string} fullpath - Fullpath to the module
-	 * @param {boolean} checkForPackaged - Whether to check for a packaged version
-	 */
-	async #loadInfoForModule(fullpath, checkForPackaged) {
-		try {
-			let isPackaged = false
-			const pkgDir = path.join(fullpath, 'pkg')
-			if (
-				checkForPackaged &&
-				(await fs.pathExists(path.join(fullpath, 'DEBUG-PACKAGED'))) &&
-				(await fs.pathExists(pkgDir))
-			) {
-				fullpath = pkgDir
-				isPackaged = true
-			}
-
-			const manifestPath = path.join(fullpath, 'companion/manifest.json')
-			if (!(await fs.pathExists(manifestPath))) {
-				this.logger.silly(`Ignoring "${fullpath}", as it is not a new module`)
-				return
-			}
-			const manifestJsonStr = await fs.readFile(manifestPath)
-			const manifestJson = JSON.parse(manifestJsonStr.toString())
-
-			validateManifest(manifestJson)
-
-			const helpPath = path.join(fullpath, 'companion/HELP.md')
-
-			const hasHelp = await fs.pathExists(helpPath)
-			const moduleDisplay = {
-				id: manifestJson.id,
-				name: manifestJson.manufacturer + ':' + manifestJson.products.join(';'),
-				version: manifestJson.version,
-				hasHelp: hasHelp,
-				bugUrl: manifestJson.bugs || manifestJson.repository,
-				shortname: manifestJson.shortname,
-				manufacturer: manifestJson.manufacturer,
-				products: manifestJson.products,
-				keywords: manifestJson.keywords,
-			}
-
-			const moduleManifestExt = {
-				manifest: manifestJson,
-				basePath: path.resolve(fullpath),
-				helpPath: hasHelp ? helpPath : null,
-				display: moduleDisplay,
-				isPackaged: isPackaged,
-			}
-
-			this.logger.silly(`found module ${moduleDisplay.id}@${moduleDisplay.version}`)
-
-			return moduleManifestExt
-		} catch (e) {
-			this.logger.silly(`Error loading module from ${fullpath}`, e)
-			this.logger.error(`Error loading module from "${fullpath}": ` + e)
-		}
 	}
 }
