+++ lib/Instance/Host.js
@@ -57,7 +57,7 @@ class ModuleHost {
 			if (!child.crashed) {
 				child.crashed = setTimeout(() => {
 					const config = this.registry.instance.store.db[child.connectionId]
-					const moduleInfo = this.registry.instance.known_modules[config?.instance_type]
+					const moduleInfo = this.registry.instance.modules.known_modules[config?.instance_type]
 
 					// Restart after a short sleep
 					this.queueRestartConnection(child.connectionId, config, moduleInfo)
@@ -105,7 +105,7 @@ class ModuleHost {
 					forceRestart()
 					return
 				}
-				// const moduleInfo = this.registry.instance.known_modules[config.instance_type]
+				// const moduleInfo = this.registry.instance.modules.known_modules[config.instance_type]
 				// if (!moduleInfo) {
 				// 	this.logger.verbose(`Missing manifest for instance "${connectionId}"`)
 				// 	forceRestart()
