+++ lib/Registry.js
@@ -255,7 +255,7 @@ class Registry extends EventEmitter {
 			process.on('message', (msg) => {
 				try {
 					if (msg.messageType === 'reload-extra-module') {
-						this.instance.reloadExtraModule(msg.fullpath).catch((e) => {
+						this.instance.modules.reloadExtraModule(msg.fullpath).catch((e) => {
 							this.logger.warn(`Failed to reload module: ${e}`)
 						})
 					}
